Add MotionEvent flag indicating that the window is obscured
This change is required to prevent tapjacking attacks.
The fix for MediaProjection vulnerability will be submitted
in separate patch. This is a security issue so it should be
merged to 5.0 and 5.1.

Contrary to the documentation FLAG_WINDOW_IS_OBSCURED
is only set when the touched point is obscured.
Changing this behaviour to reflect the documentation could
cause regressions in apps using this API. To maintain
backward compatibility the documentation has been changed
and new flag added.