Lights with Screen On [1/2]
Implement a setting allowing lights to be activated
for new notifications even if the screen is on.

Lights with screen on and Custom values are separated
in an advanced section for a cleaner overview.

This setting gives the user an oportunity to activate lights
with the screen on and also during DayDream screensaver.
The option is not activated by default.

Change-Id: I2071147d8ddab80ba0e1e7310e785ac3e03b3c7c

Conflicts:
	core/java/android/provider/Settings.java